Update code 19/10

+++ routes/api/admin.php
@@ -6,7 +6,7 @@ use App\Http\Controllers\API\Admin\AdminController;
 Route::prefix('/admin')->group(function () {
     Route::get('/profile/{id}', [AdminController::class, 'getProfile']);
     Route::post('/register', [AdminController::class, 'register']);
-    Route::post('/create-staff', [AdminController::class, 'createStaff']);
-    Route::get('/staff', [AdminController::class, 'getListStaff']);
-    Route::delete('/staff/{id}', [AdminController::class, 'deleteStaff']);
+    Route::post('/create-staff', [AdminController::class, 'createStaff'])->middleware('permission:manage staff');
+    Route::get('/staff', [AdminController::class, 'getListStaff'])->middleware('permission:manage staff');
+    Route::delete('/staff/{id}', [AdminController::class, 'deleteStaff'])->middleware('permission:manage staff');
 });
