+++ routes/api/favorite.php
@@ -3,9 +3,9 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\API\Favorite\FavoriteController;
 
-Route::prefix('/favorite')->group(function () {
-    Route::post('/', [FavoriteController::class, 'create']);
+Route::middleware(['auth', 'role:normal_user|loyal_user'])->prefix('/favorite')->group(function () {
+    Route::post('/', [FavoriteController::class, 'create'])->middleware('permission:favorite product');
     Route::get('/', [FavoriteController::class, 'get_by_user']);
     Route::get('/all', [FavoriteController::class, 'get_all']);
-    Route::delete('/{id}', [FavoriteController::class, 'delete']);
+    Route::delete('/{id}', [FavoriteController::class, 'delete'])->middleware('permission:favorite product');
 });
